Awful hack: add an intermediate grab_window for our grab.
The saver itself then will use a subwindow of the grab_window. End
result: in case the saver calls XUngrabKeyboard on the window it's told,
it no longer will stop our grab. The "bumps" XScreenSaver hack does
exactly this.

With this, I also could remove the previously existing hack to reinstate
the grab every 100ms.