Allow unauthenticated wakeup during the first second of the screen lock.
This serves to provide a quick way to prevent the screen lock from
starting when accidentally launched (e.g. while watching a video).

The amount of time is controlled by the environment variable
XSECURELOCK_ALLOW_UNAUTHENTICATED_WAKEUP_TIME_SEC and can be set in the
range from 0 (feature turned off) to 10 (maximum allowed value).

During this time, the screen saver does not run and an animation
indicating this process is drawn instead.

Requires -lrt and CLOCK_MONOTONIC.

Fixes #12.