Bring back the clamping, but instead explicitly query the root window size before building monitor structs.
The issue was that DisplayWidth()/DisplayHeight() values don't get
updated after XOpenDisplay(); the fix is querying the root window size
instead to update the auth window size.

The clamping is there to handle the (hopefully unlikely) case of a
screen being only partially in the root window (i.e. XRandR vs X11 core
inconsistency); it's no longer harmful once the bug is fixed.