Move the tritone for errors one half tone lower.
Our tonic is Emaj (thus B>E for prompt, EE for info, D#E for success);
thus the dominant seventh chord is Bmaj7. The tritone in there consists
of D# and A. Thus shall this be our tritone for errors.

To validate, play on a piano:

"prompt success": B>E<D#E (fits the chords Emaj Emaj Bmaj Emaj)
"prompt fail": B>E<AD# (fits the chords Emaj Emaj Bmaj7 Bmaj7, ends on the dominant third - harsh)

Best/worst part is that fail is usually followed by another prompt, but
this still does not sound complete, as prompt starts with B and fail
ende with a dominant third D# that really wants to go to E.