Draw auth in a black window; keep the saver visible.
This was enabled by Soumya's previous changes. Thanks!

I'd now really like to see an alternate auth helper that uses a toolkit
in single-window mode (following the monitor the mouse is on)... as
otherwise single-window auth is dead now (in fact, this version of
auth_x11 would work fine if passed the saver window ID instead).

Or maybe just a minor change to auth_x11 to behave as if there's only
one monitor all the time, while keeping the dialog on the current one...
let's see later.