New signal: SIGUSR1.
This signal is now sent from the main program to the saver to reset possible
idle timers in the savers, without intending to restart the saver.

This is part of the effort to make the auth child protocol more flexible and to
properly run auth on top of the saver as a window.