Fix memory leak in authproto.
The leak does not actually have any impact as it only occurs in the error case
in which the calling process (auth_x11) exits anyway; yet fixing this
anyway seems a good idea.

cat_authproto (the demo program) had an actual, real leak with this,
even - but it is a test program only.

Also remove a potential bad pointer dereference in case NULL isn't the
zero bit pattern.

Fixes #96.