Optionally sleep before mapping windows
This enables children some time to initialize and prepare their
content. They are started before the main loop, we sleep
XSECURELOCK_SAVER_DELAY_MS, then map the windows. The saver window is
mapped first, then the background window and the obscurer window. This
order works fine for me and I don't get a black flash anymore, despite
the saver being written in Python.

On shutdown, the order is also a bit altered to reduce the flash due
to the obscurer window in white.

The multiplexer is left untouched as its parent window is not mapped.

Fix #94