feat: enhance ui support for pam_u2f
Adds support for a interactive screen when receiving info/msg from PAM.
This means that clock will keep ticking if you receive an error or info
message from PAM. This is particularly useful for when using pam_u2f,
where you get your U2F token blinking waiting for an input while PAM
will be waiting and the screen will be left with a message "Please touch
the device". Previously that screen would seem frozen, as the clock (if
displayed) would be stopped and your screen would never go blank (if
configured). You had two options, either finish the authentication by
touching the device, or remove it so it would fail and go back to the
fallback authentication mechanism (usually user/password).

For that feature to work well, there's also a new env variable you
should set (`XSECURELOCK_AUTHPROTO_KILL_ON_TIMEOUT`): it will kill the
authproto process responsible for that auth instance. That must be done
if the PAM auth module you're using keeps waiting indefinitely for an
action from the user, which will inevitably make the screen hang forever
if that condition is not met.